+++ packages/smock/src/common/hardhat-common.ts
@@ -1,6 +1,7 @@
 /* Imports: External */
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
 import { HardhatNetworkProvider } from 'hardhat/internal/hardhat-network/provider/provider'
+import { fromHexString, toHexString } from '@eth-optimism/core-utils'
 
 /**
  * Finds the "base" Ethereum provider of the current hardhat environment.
@@ -43,4 +44,36 @@ export const findBaseHardhatProvider = (
   // TODO: Figure out a reliable way to do a type check here. Source for inspiration:
   // https://github.com/nomiclabs/hardhat/blob/master/packages/hardhat-core/src/internal/hardhat-network/provider/provider.ts
   return provider as any
+}
+
+/**
+ * Converts a string into the fancy new address thing that ethereumjs-vm v5 expects while also
+ * maintaining backwards compatibility with ethereumjs-vm v4.
+ * @param address String address to convert into the fancy new address type.
+ * @returns Fancified address.
+ */
+export const toFancyAddress = (address: string): any => {
+  const fancyAddress = fromHexString(address)
+  ;(fancyAddress as any).buf = fromHexString(address)
+  ;(fancyAddress as any).toString = (encoding?: any) => {
+    if (encoding === undefined) {
+      return address.toLowerCase()
+    } else {
+      return fromHexString(address).toString(encoding)
+    }
+  }
+  return fancyAddress
+}
+
+/**
+ * Same as toFancyAddress but in the opposite direction.
+ * @param fancyAddress Fancy address to turn into a string.
+ * @returns Way more boring address.
+ */
+export const fromFancyAddress = (fancyAddress: any): string => {
+  if (fancyAddress.buf) {
+    return toHexString(fancyAddress.buf)
+  } else {
+    return toHexString(fancyAddress)
+  }
 }